requirements.txt
# Core
python-dotenv>=1.0
pyyaml>=6.0
click>=8.1
duckdb>=0.9

# Data & ML
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
optuna>=3.4
lightgbm>=4.1
boruta>=0.3
pyarrow>=14.0

# Visualization
plotly>=5.18
matplotlib>=3.8

# Statistics
statsmodels>=0.14

# Dev
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0
tqdm>=4.66
tabulate>=0.9
joblib>=1.3
